requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
gunicorn==23.0.0
idna==3.10
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python3-openid==3.2.0
redis==5.2.1
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.9.0
